+++ CHANGELOG.md
@@ -70,13 +70,16 @@
   `@discardableResult`, etc) is on the right position:
     - If the attribute is `@IBAction` or `@NSManaged`, it should always be on
     the same line as the declaration
-    - If the attribute is `@available`, it should always be on the line above
+    - If the attribute has parameters, it should always be on the line above
     the declaration
     - Otherwise:
       - if the attribute is applied to a variable, it should be on the same line
       - if it's applied to a type or function, it should be on the line above
       - if it's applied to an import (the only option is `@testable import`), 
-      it should be on the same line.  
+      it should be on the same line.
+  You can also configure what attributes should be always in a new line or in
+  the same line as the declaration with the `always_in_same_line` and 
+  `always_in_new_line` keys.  
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#846](https://github.com/realm/SwiftLint/issues/846)